moved local serverless server into separate package
