+++ packages/cms-cli/commands/functions/test.js
@@ -1,11 +1,3 @@
-const express = require('express');
-const fs = require('fs-extra');
-const path = require('path');
-const tmp = require('tmp');
-const { spawn } = require('child_process');
-const os = require('os');
-const bodyParser = require('body-parser');
-const cors = require('cors');
 const {
   addAccountOptions,
   addConfigOptions,
@@ -21,52 +13,8 @@ const {
   checkAndWarnGitInclusion,
 } = require('@hubspot/cms-lib');
 const { logger } = require('@hubspot/cms-lib/logger');
+const { start: startTestServer } = require('local-serverless-server');
 const { validateAccount } = require('../../lib/validation');
-const defaultFunctionPackageJson = require('../../lib/templates/default-function-package.json');
-const {
-  logErrorInstance,
-} = require('@hubspot/cms-lib/errorHandlers/standardErrors');
-const { getTableContents, getTableHeader } = require('../../lib/table');
-const { processLog } = require('@hubspot/cms-lib/lib/logs');
-
-const MAX_SECRETS = 50;
-const MAX_RUNTIME = 3000;
-// AWS does not allow overriding these
-// https://docs.aws.amazon.com/lambda/latest/dg/current-supported-versions.html#lambda-environment-variables
-const AWS_RESERVED_VARS = [
-  '_HANDLER',
-  'LAMBDA_TASK_ROOT',
-  'LAMBDA_RUNTIME_DIR',
-  'AWS_EXECUTION_ENV',
-  'AWS_DEFAULT_REGION',
-  'AWS_REGION',
-  'AWS_LAMBDA_LOG_GROUP_NAME',
-  'AWS_LAMBDA_LOG_STREAM_NAME',
-  'AWS_LAMBDA_FUNCTION_NAME',
-  'AWS_LAMBDA_FUNCTION_MEMORY_SIZE',
-  'AWS_LAMBDA_FUNCTION_VERSION',
-  'AWS_ACCESS_KEY',
-  'AWS_ACCESS_KEY_ID',
-  'AWS_SECRET_KEY',
-  'AWS_SECRET_ACCESS_KEY',
-  'AWS_SESSION_TOKEN',
-  'TZ',
-];
-const AWS_RESERVED_VARS_INFO_URL =
-  'https://docs.aws.amazon.com/lambda/latest/dg/current-supported-versions.html#lambda-environment-variables';
-const DEFAULTS = {
-  HUBSPOT_LIMITS_TIME_REMAINING: 600000,
-  HUBSPOT_LIMITS_EXECUTIONS_REMAINING: 60,
-  HUBSPOT_CONTACT_VID: 123,
-  HUBSPOT_CONTACT_IS_LOGGED_IN: false,
-  HUBSPOT_CONTACT_LIST_MEMBERSHIPS: [],
-};
-
-/* TODO
-  - Make sure the shape of dataForFunc mimics shape of data passed in first param in cloud functions
-    - Determine how to pass isLoggedIn values
-  - Move to separate package?
-*/
 
 const loadAndValidateOptions = async options => {
   setLogLevel(options);
@@ -80,417 +28,26 @@ const loadAndValidateOptions = async options => {
   }
 };
 
-const installDeps = folderPath => {
-  const npmCmd = os.platform().startsWith('win') ? 'npm.cmd' : 'npm';
-  const packageJsonPath = `${folderPath}/package.json`;
-  const packageJsonExists = fs.existsSync(packageJsonPath);
-
-  if (!packageJsonExists) {
-    logger.debug(`No package.json found: using default dependencies.`);
-    fs.writeFileSync(
-      `${folderPath}/package.json`,
-      JSON.stringify(defaultFunctionPackageJson)
-    );
-  }
-
-  logger.debug(`Installing dependencies from ${folderPath}/package.json`);
-
-  return new Promise((resolve, reject) => {
-    try {
-      const npmInstallProcess = spawn(npmCmd, ['i'], {
-        env: process.env,
-        cwd: folderPath,
-      });
-
-      npmInstallProcess.on('exit', data => {
-        resolve(data);
-      });
-    } catch (e) {
-      reject(e);
-    }
-  });
-};
-
-const cleanupArtifacts = folderPath => {
-  if (fs.existsSync(folderPath)) {
-    logger.debug(`Cleaning up artifacts: ${folderPath}.`);
-    fs.rmdirSync(folderPath, { recursive: true });
-  }
-};
-
-const loadDotEnvFile = folderPath => {
-  const dotEnvPathMaybe = `${folderPath}/.env`;
-
-  if (fs.existsSync(dotEnvPathMaybe)) {
-    const loadedConfig = require('dotenv').config({ path: dotEnvPathMaybe });
-    logger.debug(`Loaded .env config from ${dotEnvPathMaybe}.`);
-    return loadedConfig;
-  }
-
-  return {};
-};
-
-const getSecrets = async (functionPath, secrets) => {
-  const config = await loadDotEnvFile(functionPath);
-  let secretsDict = {};
-
-  if (config.error) {
-    throw config.error;
-  }
-
-  secrets.forEach(secret => {
-    if (Object.prototype.hasOwnProperty.call(process.env, secret)) {
-      secretsDict[secret] = process.env[secret];
-    }
-  });
-
-  return secretsDict;
-};
-
-const loadEnvironmentVariables = (
-  globalEnvironment = {},
-  localEnvironment = {}
-) => {
-  Object.keys(globalEnvironment).forEach(globalEnvironmentVariable => {
-    if (AWS_RESERVED_VARS.indexOf(globalEnvironmentVariable) !== -1) {
-      logger.warn(
-        `The variable ${globalEnvironmentVariable} is a reserved AWS variable and should not be used. See ${AWS_RESERVED_VARS_INFO_URL} for more info.`
-      );
-    }
-
-    logger.debug(
-      `Setting environment variable(global) ${globalEnvironmentVariable} to ${localEnvironment[globalEnvironmentVariable]}.`
-    );
-    process.env[globalEnvironmentVariable] =
-      globalEnvironment[globalEnvironmentVariable];
-  });
-
-  Object.keys(localEnvironment).forEach(localEnvironmentVariable => {
-    if (AWS_RESERVED_VARS.indexOf(localEnvironmentVariable) !== -1) {
-      logger.warn(
-        `The variable ${localEnvironmentVariable} is a reserved AWS variable and should not be used. See ${AWS_RESERVED_VARS_INFO_URL} for more info.`
-      );
-    }
-
-    logger.debug(
-      `Setting environment variable(local) ${localEnvironmentVariable} to ${localEnvironment[localEnvironmentVariable]}.`
-    );
-    process.env[localEnvironmentVariable] =
-      localEnvironment[localEnvironmentVariable];
-  });
-};
-
-const getHeaders = req => {
-  const reqHeaders = req.headers;
-
-  return {
-    Accept: reqHeaders.accept,
-    'Accept-Encoding': reqHeaders['accept-encoding'],
-    'Accept-Language': reqHeaders['accept-language'],
-    'Cache-Control': reqHeaders['cache-control'],
-    Connection: reqHeaders.connection,
-    Cookie: reqHeaders.cookie,
-    Host: reqHeaders.host,
-    'True-Client-IP': req.ip, // https://stackoverflow.com/a/14631683/3612910
-    'upgrade-insecure-requests': reqHeaders['upgrade-insecure-requests'],
-    'User-Agent': reqHeaders['user-agent'],
-    'X-Forwarded-For':
-      req.headers['x-forwarded-for'] || req.connection.remoteAddress, // https://stackoverflow.com/a/14631683/3612910
-  };
-};
-
-const logFunctionExecution = (status, payload, startTime, endTime, logs) => {
-  const runTime = endTime - startTime;
-  const roundedRuntime = Math.round(runTime * 100);
-  const executionData = {
-    executionTime: runTime,
-    log: logs.join('\n'),
-    duration: `${roundedRuntime} ms`,
-    status,
-    createdAt: startTime,
-    memory: '74/128 MB',
-    id: -1,
-    payload,
-  };
-
-  logger.log(processLog(executionData, {}));
-
-  if (runTime > MAX_RUNTIME) {
-    logger.warn(
-      `Function runtime ${roundedRuntime}ms exceeded maximum runtime of ${MAX_RUNTIME}. See https://developers.hubspot.com/docs/cms/features/serverless-functions#know-your-limits for more info.`
-    );
-  }
-};
-
-const addEndpointToApp = (
-  app,
-  method,
-  route,
-  functionPath,
-  file,
-  accountId,
-  globalEnvironment,
-  localEnvironment,
-  secrets,
-  options
-) => {
-  app[method.toLowerCase()](`/${route}`, async (req, res) => {
-    const startTime = Date.now();
-    const functionFilePath = path.resolve(`${functionPath}/${file}`);
-    if (!fs.existsSync(functionFilePath)) {
-      logger.error(`Could not find file ${functionPath}/${file}.`);
-      return;
-    }
-    const { main } = require(functionFilePath);
-
-    if (!main) {
-      logger.error(`Could not find "main" export in ${functionPath}/${file}.`);
-    }
-
-    const originalConsoleLog = console.log;
-    const trackedLogs = [];
-    console.log = log => {
-      trackedLogs.push(log);
-    };
-
-    try {
-      loadEnvironmentVariables(globalEnvironment, localEnvironment);
-      const {
-        HUBSPOT_LIMITS_TIME_REMAINING,
-        HUBSPOT_LIMITS_EXECUTIONS_REMAINING,
-        HUBSPOT_CONTACT_VID,
-        HUBSPOT_CONTACT_IS_LOGGED_IN,
-        HUBSPOT_CONTACT_LIST_MEMBERSHIPS,
-      } = process.env;
-      const dataForFunc = {
-        secrets: await getSecrets(functionPath, secrets),
-        params: req.query,
-        limits: {
-          timeRemaining:
-            HUBSPOT_LIMITS_TIME_REMAINING ||
-            DEFAULTS.HUBSPOT_LIMITS_TIME_REMAINING,
-          executionsRemaining:
-            HUBSPOT_LIMITS_EXECUTIONS_REMAINING ||
-            DEFAULTS.HUBSPOT_LIMITS_EXECUTIONS_REMAINING,
-        },
-        body: req.body,
-        headers: getHeaders(req),
-        accountId,
-        contact: options.contact
-          ? {
-              vid:
-                parseInt(HUBSPOT_CONTACT_VID, 10) ||
-                DEFAULTS.HUBSPOT_CONTACT_VID,
-              isLoggedIn:
-                HUBSPOT_CONTACT_IS_LOGGED_IN ||
-                DEFAULTS.HUBSPOT_CONTACT_IS_LOGGED_IN,
-              listMemberships:
-                (HUBSPOT_CONTACT_LIST_MEMBERSHIPS.length &&
-                  HUBSPOT_CONTACT_LIST_MEMBERSHIPS.split(',')) ||
-                DEFAULTS.HUBSPOT_CONTACT_LIST_MEMBERSHIPS,
-            }
-          : null,
-      };
-
-      await main(dataForFunc, sendResponseValue => {
-        const endTime = Date.now();
-
-        console.log = originalConsoleLog;
-        logFunctionExecution(
-          'SUCCESS',
-          sendResponseValue,
-          startTime,
-          endTime,
-          trackedLogs
-        );
-        res.json(sendResponseValue);
-      });
-    } catch (e) {
-      const endTime = Date.now();
-      console.log = originalConsoleLog;
-      logFunctionExecution('UNHANDLED_ERROR', startTime, endTime, trackedLogs);
-      res.json(e);
-    }
-  });
-};
-
-const getValidatedFunctionData = functionPath => {
-  if (!fs.existsSync(functionPath)) {
-    logger.error(`The path ${functionPath} does not exist.`);
-    return;
-  } else {
-    const stats = fs.lstatSync(functionPath);
-    if (!stats.isDirectory()) {
-      logger.error(`${functionPath} is not a valid functions directory.`);
-      return;
-    }
-  }
-
-  const { endpoints, environment, secrets } = JSON.parse(
-    fs.readFileSync(`${functionPath}/serverless.json`, {
-      encoding: 'utf-8',
-    })
-  );
-  const routes = Object.keys(endpoints);
-
-  if (!routes.length) {
-    logger.error(`No endpoints found in ${functionPath}/serverless.json.`);
-    return;
-  }
-
-  if (secrets.length > MAX_SECRETS) {
-    logger.warn(
-      `This function currently exceeds the limit of ${MAX_SECRETS} secrets. See https://developers.hubspot.com/docs/cms/features/serverless-functions#know-your-limits for more info.`
-    );
-  }
-
-  return {
-    srcPath: functionPath,
-    endpoints,
-    environment,
-    routes,
-    secrets,
-  };
-};
-
-const createTemporaryFunction = async functionData => {
-  const tmpDir = tmp.dirSync();
-
-  logger.debug(`Created temporary function test folder: ${tmpDir.name}`);
-
-  await fs.copy(functionData.srcPath, tmpDir.name, {
-    overwrite: false,
-    errorOnExist: true,
-  });
-
-  await installDeps(tmpDir.name);
-
-  return {
-    ...functionData,
-    tmpDir,
-  };
-};
-
-const runTestServer = async (accountId, options) => {
-  const { path: functionPath, port } = options;
-  const validatedFunctionData = getValidatedFunctionData(functionPath);
-
-  if (!validatedFunctionData) {
-    process.exit();
-  }
-
-  const {
-    endpoints,
-    routes,
-    environment: globalEnvironment,
-    tmpDir,
-    secrets,
-  } = await createTemporaryFunction(validatedFunctionData);
-
-  const app = express();
-  app.use(bodyParser.urlencoded({ extended: true }));
-  app.use(cors());
-  app.set('trust proxy', true);
-
-  routes.forEach(route => {
-    const { method, file, environment: localEnvironment } = endpoints[route];
-
-    if (Array.isArray(method)) {
-      method.forEach(methodType => {
-        addEndpointToApp(
-          app,
-          methodType,
-          route,
-          tmpDir.name,
-          file,
-          accountId,
-          globalEnvironment,
-          localEnvironment,
-          secrets,
-          options
-        );
-      });
-    } else {
-      addEndpointToApp(
-        app,
-        method,
-        route,
-        tmpDir.name,
-        file,
-        accountId,
-        globalEnvironment,
-        localEnvironment,
-        secrets,
-        options
-      );
-    }
-  });
-
-  const localFunctionTestServer = app.listen(port, () => {
-    const testServerPath = `http://localhost:${port}`;
-    logger.log(`Local test server running at ${testServerPath}`);
-    const functionsAsArrays = routes.map(route => {
-      const { method, environment: localEnvironment } = endpoints[route];
-      return [
-        `${testServerPath}/${route}`,
-        method.join(', '),
-        secrets.join(', '),
-        Object.keys(localEnvironment).join(', '),
-      ];
-    });
-    functionsAsArrays.unshift(
-      getTableHeader([
-        'Endpoint',
-        'Methods',
-        'Secrets',
-        'Environment Variables',
-      ])
-    );
-    return logger.log(getTableContents(functionsAsArrays));
-  });
-
-  process.on('SIGINT', () => {
-    localFunctionTestServer.close();
-    logger.info('Local function test server closed.');
-    cleanupArtifacts(tmpDir.name);
-    process.exit();
-  });
-};
-
 exports.command = 'test <path>';
 exports.describe = false;
 
 exports.handler = async options => {
   loadAndValidateOptions(options);
 
-  const { path: functionPath, port } = options;
+  const { path: functionPath } = options;
   const accountId = getAccountId(options);
 
   trackCommandUsage('functions-test', { functionPath }, accountId);
-
-  const splitFunctionPath = functionPath.split('.');
-
-  if (
-    !splitFunctionPath.length ||
-    splitFunctionPath[splitFunctionPath.length - 1] !== 'functions'
-  ) {
-    logger.error(`Specified path ${functionPath} is not a .functions folder.`);
-    return;
-  }
 
   logger.debug(
     `Starting test server for .functions folder with path: ${functionPath}`
   );
 
-  try {
-    await runTestServer(accountId, options);
-  } catch (e) {
-    logErrorInstance(e, {
-      port,
-      accountId,
-      functionPath,
-    });
-  }
+  // await needed?
+  startTestServer({
+    accountId,
+    ...options,
+  });
 };
 
 exports.builder = yargs => {
