+++ packages/cms-cli/package.json
@@ -9,17 +9,14 @@
   },
   "dependencies": {
     "@hubspot/cms-lib": "^2.2.1",
+    "local-serverless-server": "^1.0.0",
     "chalk": "^4.1.0",
-    "cors": "^2.8.5",
-    "dotenv": "^8.2.0",
-    "express": "^4.17.1",
     "inquirer": "^6.3.1",
     "open": "^7.0.3",
     "ora": "^4.0.3",
     "readline": "^1.3.0",
     "shelljs": "0.8.3",
     "table": "^6.0.3",
-    "tmp": "^0.2.1",
     "update-notifier": "3.0.1",
     "yargs": "15.4.1"
   },
