+++ packages/local-serverless-server/index.js
@@ -1,0 +1,5 @@
+const { start } = require('./lib/server');
+
+module.exports = {
+  start,
+};
