+++ packages/local-serverless-server/lib/templates/default-function-package.json
