+++ packages/local-serverless-server/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "local-serverless-server",
+  "version": "1.0.0",
+  "description": "A tool for testing HubSpot CMS serverless functions locally",
+  "main": "index.js",
+  "repository": "https://github.com/HubSpot/hubspot-cms-tools",
+  "license": "Apache-2.0",
+  "dependencies": {
+    "body-parser": "^1.19.0",
+    "chalk": "^4.1.0",
+    "cors": "^2.8.5",
+    "dotenv": "^8.2.0",
+    "express": "^4.17.1",
+    "fs-extra": "^9.0.1",
+    "table": "^6.0.4",
+    "tmp": "^0.2.1"
+  }
+}
